+++ drivers/network/virtio/ethernet.h
@@ -6,10 +6,12 @@
 #pragma once
 
 #ifdef MICROKIT
-#include <microkit.h>
+#include <sys/microkit.h>
+#include <ethernet_config.h>
 #else
-#include <sel4/sel4.h>
+#include <sys/extern.h>
 #endif
+#include <sel4/sel4.h>
 #include <stdint.h>
 #include <sddf/util/printf.h>
 #include <sddf/network/queue.h>
@@ -228,4 +230,48 @@ struct resources resources;
 
 void sddf_init(void);
 void sddf_notified(uint32_t ch);
+
+#ifdef MICROKIT
+
+#define IRQ_CH 0
+#define TX_CH  1
+#define RX_CH  2
+
+uintptr_t eth_regs;
+/*
+ * The 'hardware' ring buffer region is used to store the virtIO virtqs
+ * as well as the RX and TX virtIO headers.
+ */
+uintptr_t hw_ring_buffer_vaddr;
+uintptr_t hw_ring_buffer_paddr;
+
+net_queue_t *rx_free;
+net_queue_t *rx_active;
+net_queue_t *tx_free;
+net_queue_t *tx_active;
+
+void init(void) {
+    resources = (struct resources) {
+        .regs = eth_regs,
+        .hw_ring_buffer_vaddr = hw_ring_buffer_vaddr,
+        .hw_ring_buffer_paddr = hw_ring_buffer_paddr,
+        .rx_free = rx_free,
+        .rx_active = rx_active,
+        .tx_free = tx_free,
+        .tx_active = tx_active,
+        .rx_queue_size = NET_RX_QUEUE_SIZE_DRIV,
+        .tx_queue_size = NET_TX_QUEUE_SIZE_DRIV,
+
+        .irq_id = IRQ_CH,
+        .rx_id = RX_CH,
+        .tx_id = TX_CH,
+    };
+
+    sddf_init();
+}
+
+
+#else
+
+#endif
