+++ drivers/timer/arm/timer.c
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include "sys_timer.h"
+#include "timer.h"
 #include <stdint.h>
 #include <sddf/timer/protocol.h>
 #include <sddf/util/util.h>
