+++ drivers/timer/arm/timer.h
@@ -1,5 +1,10 @@
 #pragma once
 
+#ifdef MICROKIT
+#include <sys/microkit.h>
+#else
+#include <sys/extern.h>
+#endif /* MICROKIT */
 #include <sel4/sel4.h>
 #include <stdint.h>
 
@@ -13,3 +18,17 @@ struct resources resources;
 
 void sddf_notified(unsigned int ch);
 void sddf_init();
+
+#ifdef MICROKIT
+
+#define IRQ_CH 0
+
+void init() {
+	resources = (struct resources) {
+		.irq_id = IRQ_CH,
+	};
+
+	sddf_init();
+}
+
+#endif /* MICROKIT */
