+++ examples/echo_server/board/qemu_virt_aarch64/echo_server.system
@@ -133,7 +133,7 @@
             <map mr="net_rx_buffer_data_region" vaddr="0x2_800_000" perms="r" cached="true" setvar_vaddr="virt_buffer_data_region" />
             <map mr="net_rx_buffer_data_region_cli0" vaddr="0x2_a00_000" perms="rw" cached="true" setvar_vaddr="cli_buffer_data_region" />
         </protection_domain>
-
+<!-- 
         <protection_domain name="copy1" priority="96" budget="20000" id="5">
             <program_image path="copy.elf" />
             <map mr="net_rx_free_copy1" vaddr="0x2_000_000" perms="rw" cached="true" setvar_vaddr="rx_free_virt" />
@@ -144,7 +144,7 @@
 
             <map mr="net_rx_buffer_data_region" vaddr="0x2_800_000" perms="r" cached="true" setvar_vaddr="virt_buffer_data_region" />
             <map mr="net_rx_buffer_data_region_cli1" vaddr="0x2_a00_000" perms="rw" cached="true" setvar_vaddr="cli_buffer_data_region" />
-        </protection_domain>
+        </protection_domain> -->
 
         <protection_domain name="net_virt_tx" priority="100" budget="20000" id="3">
             <program_image path="network_virt_tx.elf" />
@@ -179,7 +179,7 @@
             <map mr="cyclecounters" vaddr="0x5_010_000" perms="rw" cached="true" setvar_vaddr="cyclecounters_vaddr" />
         </protection_domain>
 
-        <protection_domain name="client1" priority="95" budget="20000" id="7">
+<!--         <protection_domain name="client1" priority="95" budget="20000" id="7">
             <program_image path="lwip.elf" />
 
             <map mr="net_rx_free_cli1" vaddr="0x2_000_000" perms="rw" cached="true" setvar_vaddr="rx_free" />
@@ -192,7 +192,7 @@
 
             <map mr="serial_tx_queue_client1" vaddr="0x4_000_000" perms="rw" cached="true" setvar_vaddr="serial_tx_queue" />
             <map mr="serial_tx_data_client1" vaddr="0x4_001_000" perms="rw" cached="true" setvar_vaddr="serial_tx_data" />
-        </protection_domain>
+        </protection_domain> -->
 
         <protection_domain name="timer" priority="101" pp="true" id="8" passive="true">
             <program_image path="timer_driver.elf" />
@@ -210,11 +210,11 @@
         <end pd="client0" id="0"/>
     </channel>
 
-    <channel>
+<!--     <channel>
         <end pd="serial_virt_tx" id="2"/>
         <end pd="client1" id="0"/>
     </channel>
-
+ -->
    <channel>
         <end pd="serial_virt_tx" id="3"/>
         <end pd="bench" id="0"/>
@@ -230,20 +230,20 @@
         <end pd="copy0" id="0" />
     </channel>
 
-    <channel>
+<!--     <channel>
         <end pd="net_virt_rx" id="2" />
         <end pd="copy1" id="0" />
-    </channel>
+    </channel> -->
 
     <channel>
         <end pd="copy0" id="1" />
         <end pd="client0" id="2" />
     </channel>
 
-    <channel>
+<!--     <channel>
         <end pd="copy1" id="1" />
         <end pd="client1" id="2" />
-    </channel>
+    </channel> -->
 
     <channel>
         <end pd="net_virt_tx" id="0" />
@@ -255,10 +255,10 @@
         <end pd="client0" id="3" />
     </channel>
 
-    <channel>
+<!--     <channel>
         <end pd="net_virt_tx" id="2" />
         <end pd="client1" id="3" />
-    </channel>
+    </channel> -->
 
     <channel>
         <end pd="client0" id="4" /> <!-- start channel -->
@@ -279,10 +279,10 @@
         <end pd="timer" id="1" />
         <end pd="client0" id="1" />
     </channel>
-
+<!-- 
     <channel>
         <end pd="timer" id="2" />
         <end pd="client1" id="1" />
-    </channel>
+    </channel> -->
 
 </system>
