+++ examples/echo_server/echo.mk
@@ -101,7 +101,7 @@ include ${SERIAL_COMPONENTS}/serial_components.mk
 
 qemu: $(IMAGE_FILE)
 	$(QEMU) -machine virt,virtualization=on \
-			-cpu cortex-a53 \
+			-cpu cortex-a57 \
 			-serial mon:stdio \
 			-device loader,file=$(IMAGE_FILE),addr=0x70000000,cpu-num=0 \
 			-m size=2G \
