+++ network/components/copy.c
@@ -3,7 +3,8 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include "sys_copy.h"
+#include "copy.h"
+#include <sddf/util/string.h>
 #include <stdbool.h>
 #include <sddf/network/queue.h>
 #include <sddf/util/string.h>
